update: Unify prisma service
