+++ backend/src/records/records.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from '../prisma.service';
+import { PrismaService } from '../prisma/prisma.service';
 import { RecordsController } from './records.controller';
 import { RecordsService } from './records.service';
